Added new methods to 3 scripts
Methods to check if wave is legal, and to legalize them into country
after a while

+++ UnityProject/Assets/Scripts/Managers/StatsManager.cs
@@ -14,7 +14,7 @@ public class StatsManager : MonoBehaviour {
 
 	double playerCurrentMoney; //because the world revolves around it
 	double playerMonthlyBudget; //how much the player will earn; depends on taxPerCitizen and BudgetBaseValue
-	double legalPopulation;//how many legal, tax paying ppl are living in your country atm
+	int legalPopulation;//how many legal, tax paying ppl are living in your country atm
 
 	//invisible variables
 	int numberOfIllegalImmigrants;
